Webasyst grouped keychain has been implemented
Keychain manager has been refactored
Tokens processing has been updated

+++ Webasyst/WatchOS/Application/WebasystApp watchOS.swift
@@ -1,10 +1,3 @@
-//
-//  WebasystApp watchOS.swift
-//  Webasyst-watchOS
-//
-//  Created by Леонид Лукашевич on 20.08.2023.
-//
-
 import Foundation
 
 public
@@ -21,15 +14,8 @@ extension WebasystApp {
     ///    - tokenType: Type of token.
     ///    - token: Token in string format for saving.
     func setToken(_ tokenType: TokenType, token: String) {
-        switch tokenType {
-        case .access:
-            let tokenData = Data(token.utf8)
-            let _ = KeychainManager.save(key: "accessToken", data: tokenData)
-            UserDefaults.standard.set(token, forKey: "accessToken")
-        case .refresh:
-            let tokenData = Data(token.utf8)
-            let _ = KeychainManager.save(key: "refreshToken", data: tokenData)
-        }
+        let tokenData = Data(token.utf8)
+        _ = KeychainManager.save(tokenType.keychainValue, data: tokenData)
     }
     
     /// Delete all records in the database
@@ -37,11 +23,16 @@ extension WebasystApp {
     func logOutUser(completion: (Bool) -> ()) {
         profileInstallService?.resetInstallList()
         profileInstallService?.deleteProfileData()
-        KeychainManager.deleteAllKeys()
-        let domain = Bundle.main.bundleIdentifier!
-        UserDefaults.standard.removePersistentDomain(forName: domain)
-        UserDefaults.standard.synchronize()
-        UserDefaults.standard.set(true, forKey: "firstLaunch")
+        
+        KeychainManager.deleteAll()
+        
+        if let domain = Bundle.main.bundleIdentifier {
+            UserDefaults.standard.removePersistentDomain(forName: domain)
+            UserDefaults.standard.synchronize()
+        }
+        
+        UserDefaults.standard.set(true, forKey: UserDefaultsKeys.firstLaunch.rawValue)
+        
         completion(true)
     }
 }
